Function scan_port moved to common module
