+++ commons/src/operator/mod.rs
@@ -1,1 +1,2 @@
+pub mod checks;
 pub mod setup;
